refactor: using duck pattern to style redux and containers
