+++ src/App.js
@@ -1,20 +1,27 @@
 import React, { Component } from 'react'
-import './App.css'
-import Header from './containers/Header'
-import Layout from './containers/Layout'
-import Sidebar from './containers/Sidebar'
-import ResultPage from './containers/ResultPage'
-import 'nodep-date-input-polyfill'
+import styled from 'styled-components'
 
+import './App.css'
+import Header from './features/fetchNews/HeaderContainer'
+import Sidebar from './features/fetchNews/SidebarContainer'
+import ResultPage from './features/fetchNews/ResultPageContainer'
+import 'nodep-date-input-polyfill'
+import { LIGHT_GREY } from './constants/colors'
+
+const StyledMainLayout = styled.main`
+  margin-top: 92px;
+  height: 100%;
+  background-color: ${LIGHT_GREY};
+`
 class App extends Component {
   render() {
     return (
       <div className="App">
         <Header />
-        <Layout>
+        <StyledMainLayout>
           <Sidebar />
           <ResultPage />
-        </Layout>
+        </StyledMainLayout>
       </div>
     )
   }
