--- src/store/actions/actionTypes.js
@@ -1,12 +1,0 @@
-export const FETCH_NEWS_INIT = 'FETCH_NEWS_INIT'
-export const FETCH_NEWS_SUCCESS = 'FETCH_NEWS_SUCCESS'
-export const FETCH_NEWS_FAIL = 'FETCH_NEWS_FAIL'
-
-export const QUERY_SEARCH = 'QUERY_SEARCH'
-export const QUERY_COUNTRY = 'QUERY_COUNTRY'
-export const QUERY_CATEGORY = 'QUERY_CATEGORY'
-export const QUERY_PAGE = 'QUERY_PAGE'
-export const QUERY = 'QUERY'
-
-export const LOADING_SHOW = 'LOADING_SHOW'
-export const LOADING_HIDE = 'LOADING_HIDE'
