--- src/store/actions/index.js
@@ -1,9 +1,0 @@
-export {
-  fetchNewsInit,
-  fetchNewsSuccess,
-  fetchNewsFail,
-  fetchNews,
-  fetchNewsHideLoading,
-} from './fetchNews'
-export { querySearch, queryCountry, queryCategory, queryPage, query } from './query'
-export { showLoading, hideLoading } from './sharedAction'
