--- src/store/actions/query.js
@@ -1,6 +1,0 @@
-import { QUERY } from './actionTypes'
-
-export const query = (key, value) => ({
-  type: QUERY,
-  payload: { [key]: value },
-})
