--- src/store/actions/sharedAction.js
@@ -1,9 +1,0 @@
-import { LOADING_HIDE, LOADING_SHOW } from './actionTypes'
-
-export const loadingShow = () => ({
-  type: LOADING_SHOW,
-})
-
-export const loadingHide = () => ({
-  type: LOADING_HIDE,
-})
